Added a prototype for another idea of a moodle question converter.
It is currently commandline based, written in `bash` and makes use
of the `curl` tool. It would be easy to create a small web frontend
which calls this script to do the conversion.

The idea is to use a full Moodle installation
to to the conversion instead of emulating Moodle. It logs
into Moodle, imports the questions given the import format.
Then it exports the questions given the export format. The
advantage of this that it is easy to support other types
than `xml` and `qtex`. A further advantage is that we do
not need to emulate Moodle. A disadvantage is the overhead
of importing it into the Moodle database, uploading the images and so on.